Improved code, removed constexpr, added exceptions
The code has been improved and constexpr-ness has been removed as it was
basically useless since nothing could be deduced at compile-time. The
PNG class has been expanded and more comments have been added. The iPNG
class has taken shape and an implementation is in the works.
Exceptions have been handled:
	- They are disabled by default, to turn them on the variable
	  PNGTRW has to be set to true (it's a global var).
	- If the exceptions are disabled, fatal errors will not throw
	  but instead they will be added to the list of errors and
	  warnings and the function using it will return the error.
	- Soft-errors, called warnings, have been implemented. These
	  will be added to the list of errors but they will not cause
	  program termination or early return of the funciton.
Things are in the works for more warnings and errors to be implemented.