Fixed namespaces, added Messages/warnings/errors
A system of messages has been added to the interface to allow the
logging of messages, warnings and errors. These all fall under the base
class PNGmsg that gives shape to how messages are like. THis class is
inherited from std::exception and contains the typical what() method
along with two new methods: type() and code() that will show the message
type and the code associated with it. Everything is structured in a
hierarchical way but this may be changed in the future if deemed
inefficient. Currently, everything has been put under the PNG namespace,
under which the major basic_PNG (formerly PNG) base class and the
derived iPNG and oPNG classes, along with all the messages classes. The
messages system is structred so that the base class contains a
std::vector of std::uniqu_ptr<PNGmsg>, which means that a queue-like
systm is implemented to allow the obtaining of the entire messages
vector, the first message and the possibility for the client to empty
the messages vector. Everything that was deemed sensitive (like the
returning of a PNG::PNGmsg derived class by the various functions) has
been handled by std::unique_ptr<PNG::PNGmsg> to allow the correct and
efficient memory freeing during implementation of the class (e.g.
correct std::vector emptying etc).
Currently the messages structure is structured like:

namespace PNG:
	-struct PNGmsg : std::exception
	-struct noMsg : PNGmsg
	namespace PNGerr:
		-struct PNGerrBase : PNGmsg
		namespace io_err:
			-struct io_file : PNGerrBase
			-struct io_stream : PNGerrBase
	namespage PNGwarn:
		-struct PNGwarn

Various other functions to better use the library have been implemented.